agents/pyproject.toml
[project]
name = "bugbounty-agents"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "web3>=7.0.0",
    "eth-abi>=5.0.0",
    "openai>=1.0.0",
    "eciespy>=0.4.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "slither-analyzer>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
